Remove -wDocumentation from Clang builds
Clang 15/16 seems to have introduced new rules for -wDocumentation with respect
to /return and /retval, which contradicts the rules we are using with Doxygen.
I am therefore removing this warning as otherwise it will break builds for
those using Clang 15/16, as the only other option is undertaking a mass
re-write of our documentation.

Signed-off-by: Paul Elliott <paul.elliott@arm.com>