Fix buffer size inconsistency
The buffer is declared with MD_MAX_SIZE, then its size is not passed to
this function, which has to know which size it was. It assumed
PSA_MAX_SIZE, but it happens that currently that's 64 in builds when
SHA-384 is enabled but not SHA-512, so larger than MD_MAX_SIZE,
resulting in stack corruption.

This is only a quickfix to get the CI to pass.
TODO: pass along the actual size of the buffer
TODO: fix PSA_HASH_MAX_SIZE too?

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>