Fix more division by zero
The previous version of the code caused a compilation error with
`gcc -Werror=vla -fundefined` due to
`PSA_ROUND_UP_TO_MULTIPLE(PSA_BLOCK_CIPHER_BLOCK_LENGTH(key_type), (input_length))`
with a block length of 0 which resulted in a division by zero, for which
UBSan inserted non-constant code in the array size. Fix this by ensuring
that in all block size calculations, the block size is nonzero.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>