Be more careful about the MAC length in PSA multipart operations
We're storing the MAC length in a uint8_t. Statically assert that this is ok.

This fixes a warning from MSVC, which is now unable to tell that the MAC
length fits in all circumstances (legitimately, it was up to the compiler's
ability to analyze a relatively complex expression in the first place).

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>