pk: keep public key in raw format
Keep public key stored into the mbedtls_pk_context structure in raw
format. In this way it is easier to get rid of several ECP/MPI functions
for parsing and writing it back. Also it fits with the format that PSA
expects for import.

Note = in some cases (es: ECSDA restartable) the old
mbedtls_ecp_keyechange structure is still used for the public key
because PSA lacks restartable support (for now)

Signed-off-by: Valerio Setti <valerio.setti@nordicsemi.no>