Refresh and fixate the semi-automatically generated code
Keep the exact output of the extraction from crypto_sizes.h, not
reformatted. Longer-than-usual lines aren't much of a problem here, but
reformatting makes it harder to spot discrepancies.

Refresh the output by running the sed command again. One line was
missing (copy-paste accident). One name was weirdly truncated, which was a
gratuitous discrepancy but had no impact on the test coverage.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>