Deprecate psa_aead_set_lengths() before setting the nonce
This possibility existed in an early draft of the PSA Crypto API
specification, but was removed before the first official release.
It's still present in Mbed TLS because we missed this change when
implementing multipart AEAD, so we now have to support it throughout
the lifetime of Mbed TLS 3.x for backward compatibility. But we intend
to remove it when we can, so announce that it's deprecated.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>