mbedtls_mpi_core_bitlen: don't support 0 limbs
Supporting edge cases such as 0 limbs is the job of the legacy bignum
module. The core module is supposed to be lean, and 0-limb arrays are
specifically documented as having undefined behavior. Move the handling of
this special case from core to legacy. The reason it was in core is that
this was one of the earliest functions moved to bignum core, before we'd
settled on where support for 0-limb bignums should live.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>