mpi_core_read: don't support 0 limbs
Supporting edge cases such as 0 limbs is the job of the legacy bignum
module. The core module is supposed to be lean, and 0-limb arrays are
specifically documented as having undefined behavior.

The reason it was in core is that this was one of the earliest code written
in core, before we'd settled on core not needing to support 0-limb arrays.

Remove the library code (saving a few bytes of code size) and corresponding
tests.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>