Weierstrass curves always have a static table
When FIXED_POINT_OPTIM == 1, all Weierstrass curves (those that use comb
multiplication) have a static table for multiples of G defined in
ecp_curve.c, and there's not reason we'd want one of them not to.

So, the code that handles the case where we don't have one, and then
fall back to computing it storing it in grp->T, is effectively dead
since 3.0.

Actually it becomes undead when adding a new curve, before it has a
static table: scripts/ecc_comb_table.py relies on this fallback kicking
in so that we can then dump the T that was computed on the fly, stick
the data in ecp_curves.c, then the fallback code becomes dead again.

Having extra complexity in the library just for supporting a development
script is rarely a good idea, especially when that complexity involves
memory management, and is the only thing that prevents us from making
the grp argument of ecp_mul const (which in turns forces us to play
tricks to ensure thread safety when an ECC PK context is shared across
threads).

So, this commit is the first in a series that will clarify how the table
of pre-computed points is used and managed.

Right now, it breaks scripts/ecc_comb_table.py, but the script will be
fixed later (by making it capable of computing the table on its own,
rather than relying on undead code in the library for that).

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>