Fix good-case testing for record decryption
We weren't actually checking that the plaintext from decryption was
identical to the original plaintext, since the backup was only shallow copy
of the record structure, and the data in buf was encrypted and decrypted in
place. Fix this by also backing up the data.

Switch buffer handling to modern helper macros. Don't modernize other macros
to keep the amount of changes small.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>