Correct Python requirements in releases
The wording wasn't quite right for 3.0.0 and up: there's nothing special
about Python and sample programs (that was true in the end times of 2.x, but
not since 3.0). Python is not needed in a release unless you want to build
the tests or you want to integrate PSA drivers without writing your own C
wrappers.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>