Add Script to add python generated files
For releases we add generated python files so users do not need python to
generate the files themselves, but on the development branch these files
are untracked and ignored by git. The script generates these files and
gets git to track them so then a release commit can be made.

Signed-off-by: Agathiyan Bragadeesh <agabra02@e127300.arm.com>