Replace scripts/ecp_comb_table.py with new program
The old script was no longer working once we removed the legacy code
that updated grp->T in ecp_mul_comb() (which was necessary to simplify
the code an in the future support making grp const.) Also, I didn't like
that it had quite a lot of C embedded in a python script.

Since the new program needs access to an MBEDTLS_STATIC_TESTABLE
function, and programs can't directly access internal header, take an
indirect route via a "test" helper, that programs can access. (This is a
bit of hack as non-test code lives in test, but I think it's OK.) Large
parts of the "test" helper are adapted from the code previously embedded
in the python script.

The new program was manually tested by running on all short Weierstrass
curves (1-8 and 10-12) and checking the output is byte for byte
identical to the current content of ecp_curves.c.

Signed-off-by: Manuel Pégourié-Gonnard <manuel.pegourie-gonnard@arm.com>