Create a base PSALowLevel class for testing low-level interfaces
As HashPSALowLevel currently has methods that can be reused in other
tests (for example, HMAC and CMAC tests), extract the commonalities
and put them under an abstract base class called PSALowLevel that
may be used in the future by other test classes.

The idea is that the classes that inherit from this base class
implement their own way of creating tests cases for a specific
algorithm, and based on this information the base class creates the
entire test suite for that algorithm.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>