pkcs7: add support for authenticated attributes
Add support for authenticated attributes as per RFC2315.

Section 9.2:
authenticatedAttributes
       [0] IMPLICIT Attributes OPTIONAL

If authenticated attributes are present, it must contain at minimum two attributes:
	- A PKCS9 content-type attribute having as its value the content type of the
	 ContentInfo value being signed.
	- A PKCS9 message-digest attribute, having as its value the message digest
	of the content.

Section 9.3:
When they are present, the result of the message digesting process is the digest
of the complete DER encoding of the section, and the expected leading tag is
SET OF and not the IMPLICIT [0] tag.

Signed-off-by: Beniamin Sandu <beniaminsandu@gmail.com>