check_test_cases/walk_ssl_opt_sh: Save the function name of a test case with params
Just identifying test cases by the 'run_tests' identifier is not
enough as the parameters that the test case is called with are also
important and should also be listed.

 * Save the description of the test case in a dictionary
   (unmodified, so that $1 is shown instead of the actual parameter
   that the function is called with) along with the function used to
   declare the test case.

For now, the saved values are not used and the relevant test cases
will not even show up in the non-executed list. This will be handled
in a future commit.

Note: To save the function names, the function declaration is
assumed to have the struction 'function_name() {' exactly

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>